+++ src/main/java/com/picksa/picksaserver/auth/exception/JwtAuthenticationException.java
@@ -1,6 +1,5 @@
 package com.picksa.picksaserver.auth.exception;
 
-import com.picksa.picksaserver.global.exception.AuthenticationException;
 import com.picksa.picksaserver.global.exception.ErrorCode;
 
 public class JwtAuthenticationException extends AuthenticationException {
