+++ src/main/java/com/picksa/picksaserver/auth/oAuth/dto/GoogleUserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.picksa.picksaserver.auth.dto;
+package com.picksa.picksaserver.auth.oAuth.dto;
 
 public class GoogleUserInfoResponse implements OAuthUserInfoResponse {
