+++ src/main/java/com/picksa/picksaserver/auth/oAuth/dto/OAuthUserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.picksa.picksaserver.auth.dto;
+package com.picksa.picksaserver.auth.oAuth.dto;
 
 public interface OAuthUserInfoResponse {
