+++ src/main/java/com/picksa/picksaserver/auth/oAuth/service/AuthCodeRequestProvider.java
@@ -1,4 +1,4 @@
-package com.picksa.picksaserver.auth.service;
+package com.picksa.picksaserver.auth.oAuth.service;
 
 public interface AuthCodeRequestProvider {
