--- src/main/java/com/picksa/picksaserver/global/exception/AuthorizationException.java
@@ -1,9 +1,0 @@
-package com.picksa.picksaserver.global.exception;
-
-public class AuthorizationException extends BusinessException {
-
-    public AuthorizationException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-
-}
